modal/requirements.txt
modal
boto3
psycopg2-binary
